WIP Refracting the `pattern` module
