--- src/patterns/mod.rs
@@ -1,5 +1,0 @@
-//! Structs to implement different types of `JoinPattern`s.
-
-pub mod binary;
-pub mod ternary;
-pub mod unary;
